requirements.txt
absl-py==0.6.1
asgiref==3.2.3
bleach==1.5.0
certifi==2019.11.28
chardet==3.0.4
distro==1.4.0
dj-database-url==0.5.0
Django==3.0
django-heroku==0.3.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
Markdown==3.0.1
numpy==1.18.0
pandas==0.25.3
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.4
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.34
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tabula-py==1.4.3
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
urllib3==1.25.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.4
